- Diminution de la fréquence de rendu. - Ajout du frustum culling sur les chunk. - Ajout du tracking de fuite mémoire. - Desactivation temporaire du flood filling. - Mise à jour de glfw. - Mise à jour de glad.

+++ src/camera.h
@@ -45,7 +45,7 @@ public:
      * @brief Met à jour la caméra.
      * @param dt La durée écoulée depuis la dernière mise à jour.
      */
-    void update(GameWindow* gl, int dt);
+    void update(GameWindow* gl, float dt);
     /**
      * @brief Fonction de mise à jour après le calcule de la physique.
      */
